requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# python version 3.12.4
# platform: osx-arm64

# conda create --name .conda python=3.12.4
# conda activate .conda
# conda install --file requirements.txt
# pip install -r requirements.txt

affine=2.4.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
appnope=0.1.4=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
bokeh=3.5.1=pypi_0
bzip2=1.0.8=h93a5062_5
ca-certificates=2024.7.4=hf0a4a13_0
cartopy=0.23.0=pypi_0
certifi=2024.7.4=pypi_0
cftime=1.6.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
cloudpickle=3.0.0=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=pypi_0
coverage=7.6.1=pypi_0
cycler=0.12.1=pypi_0
dask=2024.8.0=pypi_0
debugpy=1.8.2=py312h5c2e7bc_0
decorator=5.1.1=pyhd8ed1ab_0
deprecation=2.1.0=pypi_0
distributed=2024.8.0=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
fake-useragent=1.3.0=pypi_0
fiona=1.9.6=pypi_0
fonttools=4.53.1=pypi_0
fsspec=2024.6.1=pypi_0
gadm=0.0.5=pypi_0
geoalchemy2=0.15.2=pypi_0
geopandas=1.0.1=pypi_0
gotrue=2.6.1=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
h5netcdf=1.3.0=pypi_0
h5py=3.11.0=pypi_0
hdf4=4.2.15=h2ee6834_7
hpack=4.0.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.7=pypi_0
importlib-metadata=8.0.0=pyha770c72_0
importlib_metadata=8.0.0=hd8ed1ab_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.26.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py312h81bd7bf_0
kiwisolver=1.4.5=pypi_0
krb5=1.21.3=h237132a_0
libblas=3.9.0=22_osxarm64_openblas
libcblas=3.9.0=22_osxarm64_openblas
libcxx=18.1.8=h167917d_0
libedit=3.1.20191231=hc8eb9b7_2
libexpat=2.6.2=hebf3989_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=22_osxarm64_openblas
libopenblas=0.3.27=openmp_h517c56d_1
libsodium=1.0.18=h27ca646_1
libsqlite=3.46.0=hfb93653_0
libzlib=1.3.1=hfb2fe0b_1
line-profiler=4.1.3=pypi_0
llvm-openmp=18.1.8=hde57baf_0
locket=1.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.1=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
memory-profiler=0.61.0=pypi_0
msgpack=1.0.8=pypi_0
ncurses=6.5=hb89a1cb_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
netcdf4=1.7.1.post1=pypi_0
numpy=2.0.0=py312hb544834_0
openpyxl=3.1.5=pypi_0
openssl=3.3.1=hfb2fe0b_2
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
partd=1.4.2=pypi_0
patsy=0.5.6=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=pypi_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.2=pyhd8ed1ab_0
postgrest=0.16.9=pypi_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=6.0.0=py312h7e5086c_0
psycopg2-binary=2.9.9=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycountry=22.3.5=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyhdf=0.11.4=py312h105a5e4_1
pyogrio=0.9.0=pypi_0
pyparsing=3.1.2=pypi_0
pyproj=3.6.1=pypi_0
pyqt5=5.15.10=pypi_0
pyqt5-qt5=5.15.14=pypi_0
pyqt5-sip=12.13.0=pypi_0
pyshp=2.3.1=pypi_0
python=3.12.4=h30c5eda_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.12=4_cp312
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=py312hfa13136_0
rasterio=1.3.10=pypi_0
rasterstats=0.19.0=pypi_0
readline=8.2=h92ec313_1
realtime=1.0.6=pypi_0
requests=2.32.3=pypi_0
rioxarray=0.17.0=pypi_0
rtree=1.3.0=pypi_0
scipy=1.14.0=pypi_0
seaborn=0.13.2=pypi_0
setuptools=70.2.0=pyhd8ed1ab_0
shapely=2.0.4=pypi_0
simplejson=3.19.2=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pypi_0
snuggs=1.4.7=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlalchemy=2.0.31=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.2=pypi_0
storage3=0.7.7=pypi_0
strenum=0.4.15=pypi_0
supabase=2.6.0=pypi_0
supafunc=0.5.1=pypi_0
tblib=3.0.0=pypi_0
tk=8.6.13=h5083fa2_1
toolz=0.12.1=pypi_0
tornado=6.4.1=py312h7e5086c_0
tqdm=4.64.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
types-requests=2.32.0.20240712=pypi_0
types-setuptools=71.1.0.20240723=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.1=pypi_0
unidecode=1.3.8=pypi_0
urllib3=2.2.2=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
websockets=12.0=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
xarray=2024.6.0=pypi_0
xyzservices=2024.6.0=pypi_0
xz=5.2.6=h57fd34a_0
zeromq=4.3.5=hcc0f68c_4
zict=3.0.0=pypi_0
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.3.1=hfb2fe0b_1
